chore: move verify_results to function under IntegrationTestSimulatorConfig
commit-id:d02bdb51
